move eveything into monorepo structure
refs #164
