+++ packages/next-export-example/static/favicon.ico
