+++ packages/next-offline/index.js
