+++ packages/next-offline/service-worker.js
