+++ packages/now1-example/next.config.js
