+++ packages/now1-example/static/favicon.ico
