+++ packages/now2-example/.nowignore
