+++ packages/now2-example/next.config.js
