+++ packages/now2-example/now.json
