+++ packages/now2-example/pages/index.js
