+++ packages/now2-example/static/icon-192.png
