+++ packages/now2-example/static/icon-512.png
