+++ packages/now2-example/static/manifest.json
