+++ packages/tests/static/favicon.ico
